+++ logistic_regression/include/lr.h
@@ -21,31 +21,31 @@ public:
     //call by threads 
     void owlqn(int proc_id, int n_procs);
     //shared by multithreads
-    float *w;//model paramter shared by all threads
-    float *next_w;//model paramter after line search
-    float *global_g;//gradient of loss function
-    float *global_next_g;//gradient of loss function when arrive new w
-    float *all_nodes_global_g;
-    float global_old_loss_val;//loss value of loss function
-    float all_nodes_old_loss_val;
-    float global_new_loss_val;//loss value of loss function when arrive new w
-    float all_nodes_new_loss_val;
+    double *w;//model paramter shared by all threads
+    double *next_w;//model paramter after line search
+    double *global_g;//gradient of loss function
+    double *global_next_g;//gradient of loss function when arrive new w
+    double *all_nodes_global_g;
+    double global_old_loss_val;//loss value of loss function
+    double all_nodes_old_loss_val;
+    double global_new_loss_val;//loss value of loss function when arrive new w
+    double all_nodes_new_loss_val;
     //void* data;
     Load_Data* data;
     pid_t main_thread_id;
     int feature_dim;
-    float c;
+    double c;
     int m;
     int rank;
 //private:
-    void parallel_owlqn(int use_list_len, float* ro_list, float** s_list, float** y_list);
-    void loss_function_gradient(float *para_w, float *para_g);
-    void loss_function_subgradient(float *local_g, float *local_sub_g);
-    void two_loop(int use_list_len, float *sub_g, float **s_list, float **y_list, float *ro_list, float *p);
-    void line_search(float *local_g);
-    float loss_function_value(float *w);
-    float sigmoid(float x);
-    void fix_dir(float *w, float *next_w);
+    void parallel_owlqn(int use_list_len, double* ro_list, double** s_list, double** y_list);
+    void loss_function_gradient(double *para_w, double *para_g);
+    void loss_function_subgradient(double *local_g, double *local_sub_g);
+    void two_loop(int use_list_len, double *sub_g, double **s_list, double **y_list, double *ro_list, float *p);
+    void line_search(double *local_g);
+    double loss_function_value(double *w);
+    double sigmoid(double x);
+    void fix_dir(double *w, double *next_w);
 
     pthread_mutex_t mutex;
     pthread_barrier_t barrier;
